refactor: token_lexeme 归位到 token 模块,解开 stmt ↔ macro_system 循环
token_lexeme(&Token) -> String 是 Token 自身的属性(它就是一个大 match,
把 token 映射回源文本),却住在 macro_system 里。后果是 stmt 仅仅为了打印
一个 token 就要依赖 macro_system,而 macro_system 又要依赖 stmt 解析 for
模式——一个纯属摆放位置造成的循环依赖,挡住了两者的任何分离。

搬到 token 模块后:stmt → macro_system 的引用降为 **0**,方向变成单向
(macro_system → stmt)。

workspace 108 全绿;clippy 零警告
