+++ core/src/macro_system/proc_output.rs
@@ -1,6 +1,7 @@
-use super::{SourceToken, procedural::ProcMacroToken, token_lexeme};
+use super::{SourceToken, procedural::ProcMacroToken};
 #[cfg(not(feature = "std"))]
 use crate::compat::prelude::*;
+use crate::token::token_lexeme;
 use crate::token::{ParseError, Span, Token, Tokenizer};
 
 pub(in crate::macro_system) fn source_tokens_from_proc_output(
